Island

+++ Kitchenbuilder.Core/Models/Cooktop.cs
@@ -10,6 +10,6 @@
         // Dimensions related to cooktop placement on countertop sketch
         public int DistanceX_Cooktop_On_CT { get; set; }       // DistanceX_Cooktop@On_CT
         public int DistanceY_Cooktop_On_CT { get; set; }       // DistanceY_Cooktop@On_CT
-        public int Angle_Sketch_Rotate_Cooktop { get; set; }   // angle@Sketch_Rotate_Cooktop
+        public double Angle_Sketch_Rotate_Cooktop { get; set; }   // angle@Sketch_Rotate_Cooktop
     }
 }
